+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './collapsibleIcon/collapsibleIcon.component';
+export * from './collapsibleTitle/collapsibleTitle.component';