+++ src/decorators/index.ts
@@ -3,3 +3,4 @@ export * from './componentDisplayBlock.decorator';
 export * from './componentDisplayFlex.decorator';
 export * from './componentHostStyle.decorator';
 export * from './scrollableContent.decorator';
+export * from './typeProvider.decorator';